Create SaveFile class
